Begun work on an audio swapper class to allow for changing of voiced character lines. Swapped Gabe's non-combat lines in 3-2.
Took 1 hour 55 minutes

+++ UltrakULL/CommonFunctions.cs
@@ -25,7 +25,7 @@ namespace UltrakULL
             fadeDuration = 0.1f
         };
 
-        public static IEnumerator waitforSeconds(int seconds)
+        public static IEnumerator waitforSeconds(float seconds)
         {
             yield return new WaitForSeconds(seconds);
         }
